doc logging

+++ src/pages/api/mongo/document.js
@@ -2,8 +2,7 @@ import connectDB from '../../../helper/mongodb';
 import Document from '../../../models/Document';
 import { middleware } from "../../../middleware/middleware";
 const { createClient } = require('@supabase/supabase-js');
-const supabase = createClient(process.env.SUPABASE_URL, process.env.SUPABASE_KEY);
-  
+const supabase = createClient(process.env.SUPABASE_URL, process.env.SUPABASE_KEY);  
 
 connectDB();
 export const config = {
@@ -12,9 +11,7 @@ export const config = {
   },
 };
 
-
 export default async function handler(req, res) {
-
   const result = await middleware(req);
 
   if (!result.success) {
@@ -25,32 +22,51 @@ export default async function handler(req, res) {
     const { userRefID, docuName, type } = req.body;
 
     if (type == 'fetch') {
+      if (!userRefID) {
+        console.log("all fields not present for fetching doc")
+        return res.status(400).send('userRefID is required');
+      }
         try {
+            console.log("fetching all docs for user with user if: ", userRefID)
             const usersDocumentList = await Document.find({ userRefID: userRefID });
             if (usersDocumentList.length === 0) {
+              console.log("user has no docs")
                 return res.status(200).json({ message: 'Empty List' });
             } else {
+              console.log("user has docs and is returned")
                 return res.status(200).json(usersDocumentList);
             }
         } catch (error) {
+          console.log("theres an error", error)
             return res.status(500).json({ error: 'Error fetching users documents' });
         }
     } else if (type == 'add') {
+
+      if (!userRefID || !docuName) {
+        console.log("all fields not present forfetching doc")
+        return res.status(400).send('userRefID or docuName is not present in request');
+      }
+
+      try{
+        console.log(`adding a doc ${docuName} for user ${userRefID}`)
         const existingUserDocu = await Document.findOne({
             userRefID: userRefID,
             docuName: docuName
         });
+      
         if (existingUserDocu) {
-            return res.status(200).json({ message:'document already exists' });
-        } else {
-            try {
-                const document = new Document({userRefID, docuName });
-                await document.save();
-                return res.status(201).json({message:'Doc added success',result:document});
-            } catch (error) {
-                return res.status(500).json({ error: 'Error creating user\'s document' });
-            }
-        }
+          console.log('doc already exists')
+          return res.status(200).json({ message:'document already exists' });
+        } 
+        const document = new Document({userRefID, docuName });
+        await document.save();
+        console.log('doc saved')
+        return res.status(201).json({message:'Doc added success',result:document});
+      }catch(error){
+        console.log("there was an error", error)
+        return res.status(500).json({ error: 'Error creating user\'s document' });
+      }
+        
     } else {
         return res.status(500).json({ error: 'Wrong input body' });
     }
@@ -59,55 +75,38 @@ export default async function handler(req, res) {
    else if (req.method === 'DELETE'){
     try {
         const { docId, docName } = req.body;
-        console.log('-----------------------------')
-        console.log('deleting pages')
+
+        if (!docId) {
+          console.log("all fields not present deleting doc")
+          return res.status(400).send('docId is required');
+        }
+
+        console.log(`deleting doc with doc id :${docId}`)
+
+        console.log("deleting pages from supabase")
         const { data, error } = await supabase
         .from('pages')
         .delete()
         .match({ document: docId });
 
         if (error) {
-            console.error('Error deleting pages:', error);
+            console.error('Error deleting pages from supabase:', error);
             return res.status(500).json({ success: false, message: 'Error deleting pages' });
         } else {
-            console.log('Deleted pages:', data);
+            console.log('Deleted pages from supabase:');
         }
-        console.log('deleting document')  
+        console.log('deleting document from mongo')  
         // Delete the document
         const result = await Document.deleteMany({ docuId: docId });
         if (result.deletedCount === 0) {
             return res.status(404).json({ success: false, message: 'No documents found with the specified docId' });
           }
         console.log('mongo document deleted',result)
-        console.log('-----------------------------')
         res.status(200).json({ success: true, message: 'Document and related data deleted successfully' });
       } catch (error) {
+        console.log('there was an error',error)
         res.status(500).json({ success: false, message: 'Server error', error: error.message });
       }
     }
   }
 }
-
-/**
- * 
- -----------------------------
-deleting pinecone
-{ acknowledged: true, deletedCount: 1 }
------------------------------
-deleting Chat, only has 1 chat, if more chats
-{
-  _id: new ObjectId('65a8896fc970f13a088d285f'),
-  userID: [ 'auth0|65a4342b8dea5e613655d349' ],
-  document: 'DOC1705543978488',
-  chatName: 'what are the Ma',
-  createdAt: 2024-01-18T02:14:07.093Z,
-  updatedAt: 2024-01-18T02:14:07.093Z,
-  __v: 0
-}
-deleting chats messages
-{ acknowledged: true, deletedCount: 2 }
-deleted chats and messages
------------------------------
-deleting pages
-{ acknowledged: true, deletedCount: 22 }
- */
